refactor: split app.js into component modules
Extract 6 focused modules from the 1024-line monolithic app.js:
- state.js (36 lines) — shared mutable state
- sync-status.js (27 lines) — sync dot indicator
- alerts.js (66 lines) — RPi alerts rendering
- charts.js (466 lines) — Chart.js init, config, render, stats
- time-controls.js (122 lines) — buttons, keyboard, pan, zoom
- time-picker.js (266 lines) — Grafana-style calendar picker

app.js is now a 106-line thin orchestrator that loads data
and initializes components. All modulepreload hints added
to index.template.html for parallel fetching.

67 unit tests + 12 E2E tests pass.
