+++ gh-pages/app.js
@@ -1,59 +1,17 @@
-import {
-  lttb,
-  bucketize,
-  sortedSlice,
-  medianOf,
-  pctOf,
-  computeHistogram,
-  qualityScore,
-  rgba,
-  bsearch,
-  filterRange,
-  QUALITY_THRESHOLDS,
-  PERF_WEIGHT,
-  STAB_WEIGHT,
-  HOUR,
-  BAND_THRESHOLD,
-  PRESETS,
-  COL,
-  pad2,
-  fmtDate,
-  fmtDateTz,
-  fmtSpd,
-  fmtSpd0,
-  fmtPct,
-  histogramSVG,
-  qualityTooltipHtml,
-  statCard,
-  DS_LINE,
-  makeBandDs,
-  makeLineDs,
-} from './lib.js';
+// ── App orchestrator ─────────────────────────────────────────
+// Thin entry point: loads data, initializes components.
+// All logic lives in dedicated modules.
 
-// ── Sync status indicator ─────────────────────────────────────
-{
-  const dot = document.getElementById('syncDot');
-  const timeEl = document.querySelector('nav .meta time[datetime]');
-  if (dot && timeEl) {
-    const iso = timeEl.getAttribute('datetime');
-    const updated = new Date(iso);
-    if (!isNaN(updated)) {
-      // Allow simulating staleness via ?simAge=<minutes> (dev only)
-      const simAge = new URLSearchParams(window.location.search).get('simAge');
-      const ageMin = simAge !== null ? Number(simAge) : (Date.now() - updated.getTime()) / 60000;
-      if (ageMin < 10) {
-        dot.className = 'sync-dot sync-ok';
-        dot.title = 'Synchronisation OK (< 10 min)';
-      } else if (ageMin < 20) {
-        dot.className = 'sync-dot sync-warn';
-        dot.title = 'Synchronisation dégradée (10–20 min)';
-      } else {
-        dot.className = 'sync-dot sync-err';
-        dot.title = 'Synchronisation en échec (> 20 min)';
-      }
-    }
-  }
-}
+import { HOUR } from './lib.js';
+import { data, range, initData } from './state.js';
+import { initSyncStatus } from './sync-status.js';
+import { renderAlerts } from './alerts.js';
+import { initCharts, render, initialRender, applyZoomPlugin } from './charts.js';
+import { initTimeControls } from './time-controls.js';
+import { initTimePicker } from './time-picker.js';
+
+// ── Sync status (no data dependency) ─────────────────────────
+initSyncStatus();
 
 // ── Data loading ─────────────────────────────────────────────
 const _dataReady = Promise.all([
@@ -61,78 +19,20 @@ const _dataReady = Promise.all([
   fetch('alerts.json').then((r) => r.json()),
 ]);
 
-// ── Alerts ───────────────────────────────────────────────────
-_dataReady.then(([, ALERTS]) => {
-  // Escape HTML to prevent XSS from alert data
-  const esc = (s) =>
-    String(s).replace(
-      /[&<>"']/g,
-      (c) => ({ '&': '&amp;', '<': '&lt;', '>': '&gt;', '"': '&quot;', "'": '&#39;' })[c],
-    );
-  // Support both legacy array format and new {alerts, lastEvaluation} format
-  const alertsArr = Array.isArray(ALERTS) ? ALERTS : ALERTS?.alerts;
-  const lastEval = Array.isArray(ALERTS) ? null : ALERTS?.lastEvaluation;
-  if (!Array.isArray(alertsArr) || !alertsArr.length) {
-    document.getElementById('alertsSec').style.display = 'none';
-    return;
-  }
-  // Convert m°C to °C in alert summaries and remove duplicate thresholds
-  const fixTemp = (s) => {
-    if (!s) return '';
-    // Convert millidegrees to degrees: "70000 m°C" → "70.00 °C"
-    let r = s.replace(/(\d+)\s*m°C/g, (_, v) => (parseInt(v) / 1000).toFixed(2) + ' °C');
-    // Remove "/ 70°C" duplicate after converted threshold: "(threshold: 70.00 °C / 70°C)"
-    r = r.replace(/(\d+\.?\d*\s*°C)\s*\/\s*\d+\.?\d*\s*°C/g, '$1');
-    return r;
-  };
+// ── Alerts (independent of chart data) ───────────────────────
+_dataReady.then(([, ALERTS]) => renderAlerts(ALERTS));
 
-  // Format lastEvaluation timestamp
-  const fmtEval = (iso) => {
-    if (!iso) return null;
-    const d = new Date(iso);
-    if (isNaN(d)) return null;
-    const p = (n) => (n < 10 ? '0' + n : '' + n);
-    const tz = d.toLocaleTimeString('fr-FR', { timeZoneName: 'short' }).split(' ').pop();
-    return `${p(d.getDate())}/${p(d.getMonth() + 1)}/${d.getFullYear()} ${p(d.getHours())}:${p(
-      d.getMinutes(),
-    )}:${p(d.getSeconds())} (${tz})`;
-  };
-  const evalStr = fmtEval(lastEval);
-  const evalHtml = evalStr
-    ? `<div class="al-eval">Derni\u00e8re \u00e9valuation\u00a0: <time>${evalStr}</time></div>`
-    : '';
-
-  document.getElementById('alertsList').innerHTML =
-    evalHtml +
-    alertsArr
-      .map((a) => {
-        const icon =
-          a.state === 'firing' ? '\uD83D\uDD34' : a.state === 'pending' ? '\u26A0\uFE0F' : '\u2705';
-        const badge = a.state === 'firing' ? 'firing' : a.state === 'pending' ? 'pending' : 'ok';
-        const label = a.state === 'inactive' ? 'ok' : a.state;
-        return `<div class="al-row">
-      <span class="al-icon">${icon}</span>
-      <span class="al-name">${esc(a.name)}</span>
-      <span class="al-sum">${esc(fixTemp(a.summary))}</span>
-      <span class="al-badge ${badge}">${label}</span>
-    </div>`;
-      })
-      .join('');
-});
-
-// ── Charts & Data ────────────────────────────────────────────
+// ── Charts & controls (needs data) ──────────────────────────
 _dataReady.then(async ([RAW_DATA]) => {
-  // Yield to main thread between heavy operations to avoid long tasks (TBT)
   const yieldToMain = () => new Promise((r) => setTimeout(r, 0));
 
-  const statsEl = document.getElementById('statsRow');
-
   if (!RAW_DATA?.results?.[0]?.series?.[0]?.values?.length) {
-    statsEl.innerHTML =
+    document.getElementById('statsRow').innerHTML =
       '<div class="no-data" style="grid-column:1/-1">Aucune donn\u00e9e disponible</div>';
     return;
   }
 
+  // ── Parse InfluxDB response into typed arrays ──────────────
   const { columns: cols, values } = RAW_DATA.results[0].series[0];
   const iT = cols.indexOf('time'),
     iDl = cols.indexOf('download_bandwidth');
@@ -140,7 +40,6 @@ _dataReady.then(async ([RAW_DATA]) => {
     iPi = cols.indexOf('ping_latency');
   const LEN = values.length;
 
-  // ── Typed arrays ───────────────────────────────────────────
   const ts = new Float64Array(LEN);
   const dl = new Float64Array(LEN);
   const ul = new Float64Array(LEN);
@@ -154,171 +53,28 @@ _dataReady.then(async ([RAW_DATA]) => {
     pi[i] = r[iPi];
   }
 
+  initData(ts, dl, ul, pi);
   await yieldToMain();
 
-  const MAX_PTS = 600;
+  // ── Initialize components ──────────────────────────────────
+  await initCharts(yieldToMain);
+  await yieldToMain();
 
-  // ── Decile tooltip (click to show/hide, with bar highlight) ─
-  const dclTip = document.createElement('div');
-  dclTip.className = 'dcl-tip';
-  document.body.appendChild(dclTip);
-  let dclActiveHit = null;
+  initTimeControls();
+  initTimePicker();
+  await yieldToMain();
 
-  const dclClearSelection = () => {
-    if (dclActiveHit) {
-      const bar = dclActiveHit.nextElementSibling;
-      if (bar) bar.classList.remove('dcl-sel');
-    }
-    dclActiveHit = null;
-    dclTip.style.display = 'none';
-  };
+  // ── Initial render ─────────────────────────────────────────
+  await initialRender(yieldToMain);
+  await yieldToMain();
 
-  document.addEventListener('click', (e) => {
-    const hit = e.target.closest?.('.dcl-hit');
-    if (hit && hit !== dclActiveHit) {
-      dclClearSelection();
-      dclActiveHit = hit;
-      const bar = hit.nextElementSibling;
-      if (bar) bar.classList.add('dcl-sel');
-      dclTip.innerHTML = `<strong>${hit.dataset.tip}</strong><br><span class="dcl-tip-sub">${hit.dataset.sub}</span>`;
-      dclTip.style.display = 'block';
-      const r = hit.getBoundingClientRect();
-      dclTip.style.left = `${r.left + r.width / 2}px`;
-      dclTip.style.top = `${r.top - 4}px`;
-    } else {
-      dclClearSelection();
-    }
-  });
-
-  // ── Quality dot tooltip (click to show/hide) ───────────────
-  let _qualityData = {};
-  const qTip = document.createElement('div');
-  qTip.className = 'q-tip';
-  document.body.appendChild(qTip);
-  let qTipActive = null;
-
-  document.addEventListener('click', (e) => {
-    const dot = e.target.closest?.('.q-dot');
-    if (dot && dot !== qTipActive) {
-      qTipActive = dot;
-      const metric = dot.dataset.metric;
-      const d = _qualityData[metric];
-      if (d) {
-        qTip.innerHTML = qualityTooltipHtml(metric, d.q, d.median, d.n, d.unit);
-        qTip.style.display = 'block';
-        const r = dot.getBoundingClientRect();
-        qTip.style.left = `${r.left + r.width / 2}px`;
-        qTip.style.top = `${r.bottom + 8}px`;
-      }
-    } else if (!dot || dot === qTipActive) {
-      qTipActive = null;
-      qTip.style.display = 'none';
-    }
-  });
-
-  // ── Constants ──────────────────────────────────────────────
-  const dataEnd = ts[LEN - 1];
-  let currentH = 48,
-    rangeEnd = dataEnd + 600_000;
-  let rangeStart = rangeEnd - currentH * HOUR;
-  let isLive = true,
-    renderRAF = 0,
-    lastMode = '';
-
-  const tzAbbr = new Date().toLocaleTimeString('fr-FR', { timeZoneName: 'short' }).split(' ').pop();
-  const fmtDateTzLocal = (t) => fmtDateTz(t, tzAbbr);
-
-  const tsFilterRange = (s, e) => filterRange(ts, LEN, s, e);
-  const bucketSize = () => (currentH <= 168 ? 2 * HOUR : currentH <= 720 ? 6 * HOUR : 24 * HOUR);
-
-  // ── Gradient ───────────────────────────────────────────────
-  const mkGrad = (ctx, hex, h) => {
-    const [r, g, b] = hex
-      .slice(1)
-      .match(/.{2}/g)
-      .map((x) => parseInt(x, 16));
-    const grad = ctx.createLinearGradient(0, 0, 0, h);
-    grad.addColorStop(0, `rgba(${r},${g},${b},0.25)`);
-    grad.addColorStop(1, `rgba(${r},${g},${b},0.01)`);
-    return grad;
-  };
-
-  // ── Chart config ───────────────────────────────────────────
-  Chart.defaults.color = '#555';
-  Chart.defaults.borderColor = '#2a2a2d';
-
-  const mono = "'Geist Mono',monospace";
-  const scaleX = {
-    type: 'time',
-    time: {
-      tooltipFormat: 'dd/MM HH:mm',
-      displayFormats: { hour: 'HH:mm', day: 'dd/MM' },
-    },
-    grid: { color: '#1e2228' },
-    ticks: { font: { family: mono, size: 10 }, maxTicksLimit: 12 },
-    title: {
-      display: true,
-      text: `Heure (${tzAbbr})`,
-      font: { family: mono, size: 10 },
-      color: '#666',
-      padding: { top: 2 },
-    },
-  };
-  const baseOpts = (ar) => ({
-    responsive: true,
-    maintainAspectRatio: true,
-    aspectRatio: innerWidth < 600 ? 1.5 : ar,
-    animation: false,
-    parsing: false,
-    normalized: true,
-    interaction: { mode: 'index', intersect: false },
-    events: ['click'],
-    scales: { x: { ...scaleX } },
-    plugins: { legend: { display: false } },
-  });
-
-  // Plugin: click on visible tooltip dismisses it
-  const tooltipDismissPlugin = {
-    id: 'tooltipDismiss',
-    beforeEvent(chart, args) {
-      const evt = args.event;
-      if (evt.type !== 'click') return;
-      const tt = chart.tooltip;
-      if (!tt || !tt.opacity) return;
-      const { x, y, width, height } = tt;
-      const pad = 6;
-      if (
-        evt.x >= x - pad &&
-        evt.x <= x + width + pad &&
-        evt.y >= y - pad &&
-        evt.y <= y + height + pad
-      ) {
-        tt.setActiveElements([], { x: 0, y: 0 });
-        chart.update('none');
-        args.changed = true;
-        return false;
-      }
-    },
-  };
-  Chart.register(tooltipDismissPlugin);
-
-  const tipStyle = {
-    backgroundColor: '#141415',
-    borderColor: '#2a2a2d',
-    borderWidth: 1,
-    titleFont: { family: mono, size: 11 },
-    bodyFont: { family: mono, size: 11 },
-    padding: 10,
-    cornerRadius: 4,
-  };
-
-  // ── Drag-to-zoom (chartjs-plugin-zoom) — loaded lazily ─────
+  // ── Lazy-load drag-to-zoom plugin ──────────────────────────
   const onZoomOrPan = ({ chart }) => {
     const { min, max } = chart.scales.x;
-    rangeStart = min;
-    rangeEnd = max;
-    currentH = (rangeEnd - rangeStart) / HOUR;
-    isLive = rangeEnd >= dataEnd;
+    range.start = min;
+    range.end = max;
+    range.currentH = (range.end - range.start) / HOUR;
+    range.isLive = range.end >= range.dataEnd;
     render();
   };
   const zoomOpts = {
@@ -334,687 +90,13 @@ _dataReady.then(async ([RAW_DATA]) => {
     },
   };
 
-  // Charts created once, datasets swapped per mode
-  const bwCtx = document.getElementById('bwChart').getContext('2d');
-  const bwH = document.getElementById('bwChart').parentElement.clientHeight || 300;
-  await yieldToMain();
-
-  const bwChart = new Chart(bwCtx, {
-    type: 'line',
-    data: { datasets: [] },
-    options: {
-      ...baseOpts(2.8),
-      scales: {
-        x: { ...scaleX },
-        y: {
-          beginAtZero: true,
-          grid: { color: '#1e2228' },
-          ticks: {
-            font: { family: mono, size: 10 },
-            callback: (v) => (v >= 1000 ? (v / 1000).toFixed(1) + ' Gb/s' : v + ' Mb/s'),
-          },
-        },
-      },
-      plugins: {
-        legend: { display: false },
-        tooltip: {
-          ...tipStyle,
-          callbacks: {
-            label: (c) => {
-              const ds = c.dataset;
-              if (ds._isBand) return null; // hide band tooltip lines
-              return `${ds.label}: ${c.parsed.y.toFixed(1)} Mb/s`;
-            },
-          },
-          filter: (c) => c.dataset.label && !c.dataset._isBand,
-        },
-      },
-    },
-  });
-
-  await yieldToMain();
-
-  const piCtx = document.getElementById('piChart').getContext('2d');
-  const piH = document.getElementById('piChart').parentElement.clientHeight || 200;
-  const piChart = new Chart(piCtx, {
-    type: 'line',
-    data: { datasets: [] },
-    options: {
-      ...baseOpts(4),
-      scales: {
-        x: { ...scaleX },
-        y: {
-          beginAtZero: true,
-          grid: { color: '#1e2228' },
-          ticks: {
-            font: { family: mono, size: 10 },
-            callback: (v) => v + ' ms',
-          },
-        },
-      },
-      plugins: {
-        legend: { display: false },
-        tooltip: {
-          ...tipStyle,
-          callbacks: {
-            label: (c) => {
-              if (c.dataset._isBand) return null;
-              return `${c.parsed.y.toFixed(1)} ms`;
-            },
-          },
-          filter: (c) => !c.dataset._isBand,
-        },
-      },
-    },
-  });
-
-  // ── Band tooltip (shows bucket stats) ──────────────────────
-  const bandTooltipBw = {
-    ...tipStyle,
-    filter: (c) => !c.dataset._isBand,
-    callbacks: {
-      label: (c) => {
-        if (c.dataset._isBand) return null;
-        const val = c.parsed.y.toFixed(1);
-        return `${c.dataset.label} med: ${val} Mb/s`;
-      },
-    },
-  };
-  const bandTooltipPi = {
-    ...tipStyle,
-    filter: (c) => !c.dataset._isBand,
-    callbacks: {
-      label: (c) => (c.dataset._isBand ? null : `med: ${c.parsed.y.toFixed(1)} ms`),
-    },
-  };
-
-  // ── Render ─────────────────────────────────────────────────
-  const doRender = () => {
-    // Dismiss any active decile tooltip (stats cards are about to be rebuilt)
-    dclClearSelection();
-
-    const rng = tsFilterRange(rangeStart, rangeEnd);
-    const mode = currentH > BAND_THRESHOLD ? 'band' : 'line';
-
-    if (!rng) {
-      statsEl.innerHTML =
-        '<div class="no-data" style="grid-column:1/-1">Aucune donn\u00e9e sur cette p\u00e9riode</div>';
-      document.getElementById('bwLeg').innerHTML = '';
-      document.getElementById('piLeg').innerHTML = '';
-      bwChart.data.datasets = [];
-      piChart.data.datasets = [];
-    } else {
-      const [i0, i1] = rng;
-      const n = i1 - i0;
-
-      // Padded range: include one sample before/after for continuous curves
-      const p0 = Math.max(0, i0 - 1);
-      const p1 = Math.min(LEN, i1 + 1);
-
-      // ── Stats ──────────────────────────────────────────────
-      let dlS = 0,
-        ulS = 0,
-        piS = 0;
-      let dlMx = -Infinity,
-        dlMn = Infinity,
-        ulMx = -Infinity,
-        ulMn = Infinity;
-      let piMn = Infinity,
-        piMx = -Infinity;
-      for (let i = i0; i < i1; i++) {
-        const d = dl[i],
-          u = ul[i],
-          p = pi[i];
-        dlS += d;
-        ulS += u;
-        piS += p;
-        if (d > dlMx) dlMx = d;
-        if (d < dlMn) dlMn = d;
-        if (u > ulMx) ulMx = u;
-        if (u < ulMn) ulMn = u;
-        if (p > piMx) piMx = p;
-        if (p < piMn) piMn = p;
-      }
-      const dlAvg = dlS / n,
-        ulAvg = ulS / n,
-        piAvg = piS / n;
-      const dlSorted = sortedSlice(dl, i0, i1);
-      const ulSorted = sortedSlice(ul, i0, i1);
-      const piSorted = sortedSlice(pi, i0, i1);
-      const dlMed = medianOf(dlSorted),
-        ulMed = medianOf(ulSorted),
-        piMed = medianOf(piSorted);
-      const piP95 = pctOf(piSorted, 95);
-
-      const trLabel = `${fmtDateTzLocal(rangeStart)} \u2192 ${fmtDateTzLocal(rangeEnd)}`;
-
-      const dlHist = computeHistogram(dlSorted, dlMn, dlMx);
-      const ulHist = computeHistogram(ulSorted, ulMn, ulMx);
-      const piHist = computeHistogram(piSorted, piMn, piMx);
-
-      const dlQ = qualityScore('dl', dlMed, dlSorted, dlAvg, n);
-      const ulQ = qualityScore('ul', ulMed, ulSorted, ulAvg, n);
-      const piQ = qualityScore('pi', piMed, piSorted, piAvg, n);
-
-      // Store quality data for tooltip on click
-      _qualityData = {
-        dl: { q: dlQ, median: dlMed, n, unit: 'Mb/s' },
-        ul: { q: ulQ, median: ulMed, n, unit: 'Mb/s' },
-        pi: { q: piQ, median: piMed, n, unit: 'ms' },
-      };
-
-      statsEl.innerHTML =
-        statCard(
-          'dl',
-          'Download',
-          fmtSpd(dlMed),
-          histogramSVG(dlHist, dlAvg, dlMed, COL.dl, 'Mb/s', n),
-          `min ${fmtSpd0(dlMn)} \u00b7 avg ${fmtSpd0(dlAvg)} \u00b7 max ${fmtSpd0(dlMx)}`,
-          fmtSpd0(dl[i1 - 1]),
-          n,
-          trLabel,
-          dlQ,
-        ) +
-        statCard(
-          'ul',
-          'Upload',
-          fmtSpd(ulMed),
-          histogramSVG(ulHist, ulAvg, ulMed, COL.ul, 'Mb/s', n),
-          `min ${fmtSpd0(ulMn)} \u00b7 avg ${fmtSpd0(ulAvg)} \u00b7 max ${fmtSpd0(ulMx)}`,
-          fmtSpd0(ul[i1 - 1]),
-          n,
-          trLabel,
-          ulQ,
-        ) +
-        statCard(
-          'pi',
-          'Ping',
-          `${piMed.toFixed(1)}<span class="u">ms</span>`,
-          histogramSVG(piHist, piAvg, piMed, COL.pi, 'ms', n),
-          `min ${piMn.toFixed(1)} \u00b7 avg ${piAvg.toFixed(1)} \u00b7 max ${piMx.toFixed(1)}`,
-          pi[i1 - 1].toFixed(1),
-          n,
-          trLabel,
-          piQ,
-        );
-
-      // ── Legend ─────────────────────────────────────────────
-      const bandLabel =
-        mode === 'band' ? ' <span style="color:var(--text3)">(m\u00e9diane + IQR)</span>' : '';
-      document.getElementById('bwLeg').innerHTML = `
-        <span><i style="background:${COL.dl}"></i>download${bandLabel}</span>
-        <span><i style="background:${COL.ul}"></i>upload</span>`;
-      document.getElementById('piLeg').innerHTML = `
-        <span><i style="background:${COL.pi}"></i>latency${bandLabel}</span>`;
-
-      // ── Datasets (padded range for edge continuity) ────────
-      if (mode === 'band') {
-        const bMs = bucketSize();
-        const dlB = bucketize(ts, dl, p0, p1, bMs);
-        const ulB = bucketize(ts, ul, p0, p1, bMs);
-        const piB = bucketize(ts, pi, p0, p1, bMs);
-
-        bwChart.data.datasets = [
-          ...makeBandDs(dlB, COL.dl, 'download'),
-          ...makeBandDs(ulB, COL.ul, 'upload'),
-        ];
-        bwChart.options.plugins.tooltip = bandTooltipBw;
-
-        piChart.data.datasets = makeBandDs(piB, COL.pi, 'latency');
-        piChart.options.plugins.tooltip = bandTooltipPi;
-      } else {
-        bwChart.data.datasets = [
-          ...makeLineDs(ts, dl, p0, p1, COL.dl, 'download', mkGrad(bwCtx, COL.dl, bwH)),
-          ...makeLineDs(ts, ul, p0, p1, COL.ul, 'upload', mkGrad(bwCtx, COL.ul, bwH)),
-        ];
-        bwChart.options.plugins.tooltip = {
-          ...tipStyle,
-          callbacks: {
-            label: (c) => `${c.dataset.label}: ${c.parsed.y.toFixed(1)} Mb/s`,
-          },
-        };
-
-        piChart.data.datasets = makeLineDs(
-          ts,
-          pi,
-          p0,
-          p1,
-          COL.pi,
-          'latency',
-          mkGrad(piCtx, COL.pi, piH),
-        );
-        piChart.options.plugins.tooltip = {
-          ...tipStyle,
-          callbacks: {
-            label: (c) => `${c.parsed.y.toFixed(1)} ms`,
-          },
-        };
-      }
-    }
-
-    bwChart.options.scales.x.min = rangeStart;
-    bwChart.options.scales.x.max = rangeEnd;
-
-    piChart.options.scales.x.min = rangeStart;
-    piChart.options.scales.x.max = rangeEnd;
-
-    document.getElementById('rangeLabel').textContent = `${fmtDateTzLocal(
-      rangeStart,
-    )}  \u2192  ${fmtDateTzLocal(rangeEnd)}`;
-    document
-      .querySelectorAll('.rb')
-      .forEach((b) => b.classList.toggle('on', parseInt(b.dataset.hours) === currentH));
-    lastMode = mode;
-  };
-
-  const flushCharts = () => {
-    bwChart.update('none');
-    piChart.update('none');
-  };
-
-  const render = () => {
-    cancelAnimationFrame(renderRAF);
-    renderRAF = requestAnimationFrame(() => {
-      doRender();
-      flushCharts();
-    });
-  };
-
-  // ── Time controls ──────────────────────────────────────────
-  const setRange = (h) => {
-    currentH = h;
-    if (isLive) rangeEnd = dataEnd + 600_000;
-    rangeStart = rangeEnd - currentH * HOUR;
-    render();
-  };
-
-  const shift = (dir) => {
-    const s = currentH * HOUR * 0.5 * dir;
-    rangeStart += s;
-    rangeEnd += s;
-    isLive = rangeEnd >= dataEnd;
-    render();
-  };
-
-  document.querySelectorAll('.rb').forEach((b) =>
-    b.addEventListener('click', () => {
-      isLive = true;
-      setRange(parseInt(b.dataset.hours));
-    }),
-  );
-  document.getElementById('btnBack').addEventListener('click', () => shift(-1));
-  document.getElementById('btnFwd').addEventListener('click', () => shift(1));
-  document.getElementById('btnZoomIn').addEventListener('click', () => {
-    const i = PRESETS.indexOf(currentH);
-    if (i > 0) setRange(PRESETS[i - 1]);
-    else if (currentH > PRESETS[0]) setRange(PRESETS[0]);
-  });
-  document.getElementById('btnZoomOut').addEventListener('click', () => {
-    const i = PRESETS.indexOf(currentH);
-    if (i >= 0 && i < PRESETS.length - 1) setRange(PRESETS[i + 1]);
-    else if (i < 0) {
-      for (const p of PRESETS) {
-        if (p > currentH) {
-          setRange(p);
-          break;
-        }
-      }
-    }
-  });
-  document.addEventListener('keydown', (e) => {
-    if (e.key === 'ArrowLeft') shift(-1);
-    else if (e.key === 'ArrowRight') shift(1);
-    else if (e.key === '+' || e.key === '=') document.getElementById('btnZoomIn').click();
-    else if (e.key === '-') document.getElementById('btnZoomOut').click();
-  });
-
-  // Double-click on chart → reset to default 48 h live view
-  const resetToLive = () => {
-    isLive = true;
-    setRange(48);
-  };
-  document.getElementById('bwChart').addEventListener('dblclick', resetToLive);
-  document.getElementById('piChart').addEventListener('dblclick', resetToLive);
-
-  // ── Middle-click drag to pan ───────────────────────────────
-  ['bwChart', 'piChart'].forEach((id) => {
-    const canvas = document.getElementById(id);
-    let panning = false,
-      startX = 0,
-      startRange = 0;
-    canvas.addEventListener('mousedown', (e) => {
-      if (e.button !== 1) return; // middle button only
-      e.preventDefault();
-      panning = true;
-      startX = e.clientX;
-      startRange = rangeStart;
-      canvas.style.cursor = 'grabbing';
-    });
-    window.addEventListener('mousemove', (e) => {
-      if (!panning) return;
-      const chart = Chart.getChart(id);
-      const xScale = chart.scales.x;
-      const pxRange = xScale.right - xScale.left;
-      const msPerPx = (rangeEnd - rangeStart) / pxRange;
-      const dx = startX - e.clientX; // drag left → move range forward
-      const shift = dx * msPerPx;
-      rangeStart = startRange + shift;
-      rangeEnd = rangeStart + currentH * HOUR;
-      isLive = rangeEnd >= dataEnd;
-      render();
-    });
-    const stopPan = () => {
-      if (!panning) return;
-      panning = false;
-      canvas.style.cursor = 'crosshair';
-    };
-    window.addEventListener('mouseup', (e) => {
-      if (e.button === 1) stopPan();
-    });
-    // Prevent default middle-click scroll behavior
-    canvas.addEventListener('auxclick', (e) => {
-      if (e.button === 1) e.preventDefault();
-    });
-  });
-
-  await yieldToMain();
-
-  // ── Grafana-style time range picker ────────────────────────
-  (() => {
-    const picker = document.getElementById('trPicker');
-    const btn = document.getElementById('rangeLabelBtn');
-    const calEl = document.getElementById('trCal');
-    const fromIn = document.getElementById('trFrom');
-    const toIn = document.getElementById('trTo');
-    const recentSec = document.getElementById('trRecentSec');
-    const recentEl = document.getElementById('trRecent');
-    const relEl = document.getElementById('trRel');
-
-    let pickerOpen = false;
-    let calYear, calMonth; // currently displayed month
-    let selFrom = null,
-      selTo = null; // calendar selection state (timestamps)
-    const dataStart = ts[0];
-
-    // Relative presets (label, hours)
-    const relPresets = [
-      ['5 min', 5 / 60],
-      ['15 min', 0.25],
-      ['30 min', 0.5],
-      ['1 heure', 1],
-      ['3 heures', 3],
-      ['6 heures', 6],
-      ['12 heures', 12],
-      ['24 heures', 24],
-      ['2 jours', 48],
-      ['7 jours', 168],
-      ['30 jours', 720],
-    ];
-
-    // Recent ranges (stored in memory, max 5)
-    let recentRanges = [];
-    const addRecent = (s, e) => {
-      const key = `${s}|${e}`;
-      recentRanges = recentRanges.filter((r) => `${r[0]}|${r[1]}` !== key);
-      recentRanges.unshift([s, e]);
-      if (recentRanges.length > 5) recentRanges.length = 5;
-      renderRecent();
-    };
-
-    // ── Toggle ──
-    const togglePicker = () => {
-      pickerOpen = !pickerOpen;
-      picker.style.display = pickerOpen ? '' : 'none';
-      btn.classList.toggle('open', pickerOpen);
-      if (pickerOpen) syncPickerToRange();
-    };
-    btn.addEventListener('click', togglePicker);
-
-    // Close on outside click
-    document.addEventListener('click', (e) => {
-      if (pickerOpen && !picker.contains(e.target) && !btn.contains(e.target)) {
-        pickerOpen = false;
-        picker.style.display = 'none';
-        btn.classList.remove('open');
-      }
-    });
-
-    // ── Sync picker inputs to current range ──
-    const toLocalISO = (t) => {
-      const d = new Date(t);
-      const off = d.getTimezoneOffset();
-      const local = new Date(d.getTime() - off * 60000);
-      return local.toISOString().slice(0, 16);
-    };
-    const syncPickerToRange = () => {
-      selFrom = rangeStart;
-      selTo = rangeEnd;
-      fromIn.value = toLocalISO(rangeStart);
-      toIn.value = toLocalISO(rangeEnd);
-      const d = new Date(rangeStart);
-      calYear = d.getFullYear();
-      calMonth = d.getMonth();
-      renderCal();
-      renderRelActive();
-    };
-
-    // ── Apply absolute range ──
-    const applyAbsolute = () => {
-      const s = new Date(fromIn.value).getTime();
-      const e = new Date(toIn.value).getTime();
-      if (!s || !e || s >= e) return;
-      rangeStart = s;
-      rangeEnd = e;
-      currentH = (rangeEnd - rangeStart) / HOUR;
-      isLive = rangeEnd >= dataEnd;
-      addRecent(s, e);
-      render();
-      togglePicker();
-    };
-    document.getElementById('trApply').addEventListener('click', applyAbsolute);
-
-    // Sync calendar highlight when inputs change manually
-    fromIn.addEventListener('change', () => {
-      const t = new Date(fromIn.value).getTime();
-      if (t) selFrom = t;
-      if (toIn.value) selTo = new Date(toIn.value).getTime();
-      renderCal();
-    });
-    toIn.addEventListener('change', () => {
-      if (fromIn.value) selFrom = new Date(fromIn.value).getTime();
-      const t = new Date(toIn.value).getTime();
-      if (t) selTo = t;
-      renderCal();
-    });
-
-    // ── Relative presets ──
-    relPresets.forEach(([label, h]) => {
-      const b = document.createElement('button');
-      b.textContent = label;
-      b.dataset.hours = h;
-      b.addEventListener('click', () => {
-        isLive = true;
-        currentH = h;
-        rangeEnd = dataEnd + 600_000;
-        rangeStart = rangeEnd - currentH * HOUR;
-        render();
-        togglePicker();
-      });
-      relEl.appendChild(b);
-    });
-    const renderRelActive = () => {
-      relEl.querySelectorAll('button').forEach((b) => {
-        const h = parseFloat(b.dataset.hours);
-        b.classList.toggle('active', Math.abs(h - currentH) < 0.01 && isLive);
-      });
-    };
-
-    // ── Calendar ──
-    const DAYS = ['Lu', 'Ma', 'Me', 'Je', 'Ve', 'Sa', 'Di'];
-    const MONTHS = [
-      'Janvier',
-      'F\u00e9vrier',
-      'Mars',
-      'Avril',
-      'Mai',
-      'Juin',
-      'Juillet',
-      'Ao\u00fbt',
-      'Septembre',
-      'Octobre',
-      'Novembre',
-      'D\u00e9cembre',
-    ];
-    const renderCal = () => {
-      const first = new Date(calYear, calMonth, 1);
-      const startDay = (first.getDay() + 6) % 7; // Mon=0
-      const daysInMonth = new Date(calYear, calMonth + 1, 0).getDate();
-      const prevDays = new Date(calYear, calMonth, 0).getDate();
-      const today = new Date();
-      const todayStr = `${today.getFullYear()}-${today.getMonth()}-${today.getDate()}`;
-
-      // Use picker selection (from inputs) for calendar highlighting
-      const hFrom = selFrom || 0;
-      const hTo = selTo || 0;
-
-      let html = `<div class="tr-cal-hd">
-        <button id="trCalPrev">&lsaquo;</button>
-        <span>${MONTHS[calMonth]} ${calYear}</span>
-        <button id="trCalNext">&rsaquo;</button>
-      </div><table><tr>${DAYS.map((d) => `<th>${d}</th>`).join('')}</tr><tr>`;
-
-      let cell = 0;
-      // Previous month padding
-      for (let i = startDay - 1; i >= 0; i--) {
-        const day = prevDays - i;
-        html += `<td><button class="other">${day}</button></td>`;
-        cell++;
-      }
-      // Current month
-      for (let d = 1; d <= daysInMonth; d++) {
-        if (cell && cell % 7 === 0) html += '</tr><tr>';
-        const dt = new Date(calYear, calMonth, d);
-        const dayStart = dt.getTime();
-        const dayEnd = dayStart + 86400000;
-        const str = `${calYear}-${calMonth}-${d}`;
-        const cls = [];
-        if (str === todayStr) cls.push('today');
-        if (dayEnd < dataStart || dayStart > dataEnd + 600_000) cls.push('no-data');
-        else if (hFrom && hTo) {
-          // Highlight based on picker From/To selection
-          const rS = Math.min(hFrom, hTo);
-          const rE = Math.max(hFrom, hTo);
-          if (dayStart >= rS && dayEnd <= rE) cls.push('in-range');
-          if ((dayStart <= rS && dayEnd > rS) || (dayStart < rE && dayEnd >= rE)) cls.push('sel');
-        }
-        html += `<td><button class="${cls.join(' ')}" data-ts="${dayStart}">${d}</button></td>`;
-        cell++;
-      }
-      // Next month padding
-      let nd = 1;
-      while (cell % 7 !== 0) {
-        html += `<td><button class="other">${nd++}</button></td>`;
-        cell++;
-      }
-      html += '</tr></table>';
-      calEl.innerHTML = html;
-
-      // Calendar nav
-      document.getElementById('trCalPrev').addEventListener('click', (e) => {
-        e.stopPropagation();
-        calMonth--;
-        if (calMonth < 0) {
-          calMonth = 11;
-          calYear--;
-        }
-        renderCal();
-      });
-      document.getElementById('trCalNext').addEventListener('click', (e) => {
-        e.stopPropagation();
-        calMonth++;
-        if (calMonth > 11) {
-          calMonth = 0;
-          calYear++;
-        }
-        renderCal();
-      });
-
-      // Day click: first click → set From, second click → set To
-      calEl.querySelectorAll('td button[data-ts]').forEach((b) => {
-        if (b.classList.contains('no-data')) return;
-        b.addEventListener('click', (e) => {
-          e.stopPropagation();
-          const t = parseInt(b.dataset.ts);
-          if (!selFrom || selTo) {
-            // First click or reset: set From
-            selFrom = t;
-            selTo = null;
-            fromIn.value = toLocalISO(t);
-            toIn.value = '';
-          } else {
-            // Second click: set To (ensure From < To)
-            const a = Math.min(selFrom, t);
-            const z = Math.max(selFrom, t) + 86400000; // end of the later day
-            selFrom = a;
-            selTo = z;
-            fromIn.value = toLocalISO(a);
-            toIn.value = toLocalISO(z);
-          }
-          renderCal();
-        });
-      });
-    };
-
-    // ── Recent ranges ──
-    const renderRecent = () => {
-      if (!recentRanges.length) {
-        recentSec.style.display = 'none';
-        return;
-      }
-      recentSec.style.display = '';
-      recentEl.innerHTML = recentRanges
-        .map(
-          ([s, e]) =>
-            `<button data-s="${s}" data-e="${e}">${fmtDate(s)} \u2192 ${fmtDate(e)}</button>`,
-        )
-        .join('');
-      recentEl.querySelectorAll('button').forEach((b) =>
-        b.addEventListener('click', () => {
-          rangeStart = parseInt(b.dataset.s);
-          rangeEnd = parseInt(b.dataset.e);
-          currentH = (rangeEnd - rangeStart) / HOUR;
-          isLive = rangeEnd >= dataEnd;
-          render();
-          togglePicker();
-        }),
-      );
-    };
-  })();
-
-  await yieldToMain();
-
-  // Initial render — split chart updates to avoid long tasks
-  doRender();
-  bwChart.update('none');
-  await yieldToMain();
-  piChart.update('none');
-  await yieldToMain();
-
-  // ── Lazy-load hammerjs + chartjs-plugin-zoom after first paint ──
   const loadZoom = () => {
     const s1 = document.createElement('script');
     s1.src = 'https://cdn.jsdelivr.net/npm/hammerjs@2';
     s1.onload = () => {
       const s2 = document.createElement('script');
       s2.src = 'https://cdn.jsdelivr.net/npm/chartjs-plugin-zoom@2';
-      s2.onload = () => {
-        bwChart.options.plugins.zoom = zoomOpts;
-        bwChart.update('none');
-        setTimeout(() => {
-          piChart.options.plugins.zoom = zoomOpts;
-          piChart.update('none');
-        }, 0);
-      };
+      s2.onload = () => applyZoomPlugin(zoomOpts);
       document.head.appendChild(s2);
     };
     document.head.appendChild(s1);
