+++ gh-pages/index.template.html
@@ -15,6 +15,12 @@
     <link rel="preload" href="https://cdn.jsdelivr.net/npm/chart.js@4" as="script" />
     <link rel="preload" href="fonts/geist-latin.woff2" as="font" type="font/woff2" crossorigin />
     <link rel="modulepreload" href="lib.js" />
+    <link rel="modulepreload" href="state.js" />
+    <link rel="modulepreload" href="sync-status.js" />
+    <link rel="modulepreload" href="alerts.js" />
+    <link rel="modulepreload" href="charts.js" />
+    <link rel="modulepreload" href="time-controls.js" />
+    <link rel="modulepreload" href="time-picker.js" />
     <link rel="stylesheet" href="style.css" />
   </head>
   <body>
